add example for fan-in pattern
